v3/requirements.txt
pandas
requests
zipfile36
psycopg2-binary
tabulate
spacy
rapidfuzz
sentence-transformers
torch